Extract Scheme-related structs and enums into their own files
