+++ Sources/ProjectDescription/Arguments.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+public struct Arguments: Equatable, Codable {
+    public let environment: [String: String]
+    public let launch: [String: Bool]
+
+    public init(environment: [String: String] = [:],
+                launch: [String: Bool] = [:]) {
+        self.environment = environment
+        self.launch = launch
+    }
+}
