+++ Sources/ProjectDescription/Scheme.swift
@@ -24,176 +24,3 @@ public struct Scheme: Equatable, Codable {
         self.archiveAction = archiveAction
     }
 }
-
-// MARK: - ExecutionAction
-
-public struct ExecutionAction: Equatable, Codable {
-    public let title: String
-    public let scriptText: String
-    public let target: TargetReference?
-
-    public init(title: String = "Run Script", scriptText: String, target: TargetReference? = nil) {
-        self.title = title
-        self.scriptText = scriptText
-        self.target = target
-    }
-}
-
-// MARK: - Arguments
-
-public struct Arguments: Equatable, Codable {
-    public let environment: [String: String]
-    public let launch: [String: Bool]
-
-    public init(environment: [String: String] = [:],
-                launch: [String: Bool] = [:]) {
-        self.environment = environment
-        self.launch = launch
-    }
-}
-
-// MARK: - BuildAction
-
-public struct BuildAction: Equatable, Codable {
-    public let targets: [TargetReference]
-    public let preActions: [ExecutionAction]
-    public let postActions: [ExecutionAction]
-
-    public init(targets: [TargetReference],
-                preActions: [ExecutionAction] = [],
-                postActions: [ExecutionAction] = []) {
-        self.targets = targets
-        self.preActions = preActions
-        self.postActions = postActions
-    }
-}
-
-// MARK: - TestAction
-
-public struct TestAction: Equatable, Codable {
-    public let targets: [TestableTarget]
-    public let arguments: Arguments?
-    public let configurationName: String
-    public let coverage: Bool
-    public let codeCoverageTargets: [TargetReference]
-    public let preActions: [ExecutionAction]
-    public let postActions: [ExecutionAction]
-
-    public init(targets: [TestableTarget] = [],
-                arguments: Arguments? = nil,
-                configurationName: String,
-                coverage: Bool = false,
-                codeCoverageTargets: [TargetReference] = [],
-                preActions: [ExecutionAction] = [],
-                postActions: [ExecutionAction] = []) {
-        self.targets = targets
-        self.arguments = arguments
-        self.configurationName = configurationName
-        self.coverage = coverage
-        self.preActions = preActions
-        self.postActions = postActions
-        self.codeCoverageTargets = codeCoverageTargets
-    }
-
-    public init(targets: [TestableTarget],
-                arguments: Arguments? = nil,
-                config: PresetBuildConfiguration = .debug,
-                coverage: Bool = false,
-                codeCoverageTargets: [TargetReference] = [],
-                preActions: [ExecutionAction] = [],
-                postActions: [ExecutionAction] = []) {
-        self.init(targets: targets,
-                  arguments: arguments,
-                  configurationName: config.name,
-                  coverage: coverage,
-                  codeCoverageTargets: codeCoverageTargets,
-                  preActions: preActions,
-                  postActions: postActions)
-    }
-}
-
-public struct TestableTarget: Equatable, Codable, ExpressibleByStringLiteral {
-    public let target: TargetReference
-    public let isSkipped: Bool
-    public let isParallelizable: Bool
-    public let isRandomExecutionOrdering: Bool
-
-    public init(target: TargetReference, skipped: Bool = false, parallelizable: Bool = false, randomExecutionOrdering: Bool = false) {
-        self.target = target
-        isSkipped = skipped
-        isParallelizable = parallelizable
-        isRandomExecutionOrdering = randomExecutionOrdering
-    }
-
-    public init(stringLiteral value: String) {
-        self.init(target: .init(projectPath: nil, target: value))
-    }
-}
-
-// MARK: - RunAction
-
-public struct RunAction: Equatable, Codable {
-    public let configurationName: String
-    public let executable: TargetReference?
-    public let arguments: Arguments?
-
-    public init(configurationName: String,
-                executable: TargetReference? = nil,
-                arguments: Arguments? = nil) {
-        self.configurationName = configurationName
-        self.executable = executable
-        self.arguments = arguments
-    }
-
-    public init(config: PresetBuildConfiguration = .debug,
-                executable: TargetReference? = nil,
-                arguments: Arguments? = nil) {
-        self.init(configurationName: config.name,
-                  executable: executable,
-                  arguments: arguments)
-    }
-}
-
-// MARK: - Archive Action
-
-public struct ArchiveAction: Equatable, Codable {
-    public let configurationName: String
-    public let revealArchiveInOrganizer: Bool
-    public let customArchiveName: String?
-    public let preActions: [ExecutionAction]
-    public let postActions: [ExecutionAction]
-
-    public init(
-        configurationName: String,
-        revealArchiveInOrganizer: Bool = true,
-        customArchiveName: String? = nil,
-        preActions: [ExecutionAction] = [],
-        postActions: [ExecutionAction] = []
-    ) {
-        self.configurationName = configurationName
-        self.revealArchiveInOrganizer = revealArchiveInOrganizer
-        self.customArchiveName = customArchiveName
-        self.preActions = preActions
-        self.postActions = postActions
-    }
-}
-
-// MARK: - Target Reference
-
-public struct TargetReference: Equatable, Codable, ExpressibleByStringLiteral {
-    public var projectPath: Path?
-    public var targetName: String
-
-    public init(projectPath: Path?, target: String) {
-        self.projectPath = projectPath
-        targetName = target
-    }
-
-    public init(stringLiteral value: String) {
-        self = .init(projectPath: nil, target: value)
-    }
-
-    public static func project(path: Path, target: String) -> TargetReference {
-        .init(projectPath: path, target: target)
-    }
-}
