+++ Sources/TuistCore/Models/RunAction.swift
@@ -1,0 +1,28 @@
+import Basic
+import Foundation
+
+public struct RunAction: Equatable {
+    // MARK: - Attributes
+
+    public let configurationName: String
+    public let executable: TargetReference?
+    public let arguments: Arguments?
+
+    // MARK: - Init
+
+    public init(configurationName: String,
+                executable: TargetReference? = nil,
+                arguments: Arguments? = nil) {
+        self.configurationName = configurationName
+        self.executable = executable
+        self.arguments = arguments
+    }
+
+    // MARK: - Equatable
+
+    public static func == (lhs: RunAction, rhs: RunAction) -> Bool {
+        lhs.configurationName == rhs.configurationName &&
+            lhs.executable == rhs.executable &&
+            lhs.arguments == rhs.arguments
+    }
+}
