+++ Sources/TuistCoreTesting/Models/Arguments+TestData.swift
@@ -1,0 +1,11 @@
+import Basic
+import Foundation
+@testable import TuistCore
+
+public extension Arguments {
+    static func test(environment: [String: String] = [:],
+                     launch: [String: Bool] = [:]) -> Arguments {
+        Arguments(environment: environment,
+                  launch: launch)
+    }
+}
