+++ Sources/TuistCoreTesting/Models/BuildAction+TestData.swift
@@ -1,0 +1,11 @@
+import Basic
+import Foundation
+@testable import TuistCore
+
+public extension BuildAction {
+    static func test(targets: [TargetReference] = [TargetReference(projectPath: "/Project", name: "App")],
+                     preActions: [ExecutionAction] = [],
+                     postActions: [ExecutionAction] = []) -> BuildAction {
+        BuildAction(targets: targets, preActions: preActions, postActions: postActions)
+    }
+}
